requirements.txt
annotated-types==0.7.0
anyio==4.8.0
click==8.1.8
fastapi==0.115.8
h11==0.14.0
idna==3.10
Jinja2==3.1.5
MarkupSafe==3.0.2
pyaes==1.6.1
pyasn1==0.6.1
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.0.1
rsa==4.9
sniffio==1.3.1
starlette==0.45.3
Telethon==1.38.1
typing_extensions==4.12.2
uvicorn==0.34.0
